[code] moved function to separate module
